Working decoding the joystick on ESP32 :-)
Reading the Linux Kernel driver and adapting some of its code was of great help despite some things are handled in a strange way.
The Linux Code seems to be meant to poll the Gameport, but we mostly use GPIO interrupts now (both edges). In the future we might
be able to more cleverly use interrupts instead of the way the kernel driver compares current and previous port values.

Thanks a lot to Vojtech Pavlik who wrote the kernel driver more than 20 years ago!